Fix the composer provide rules
This package does not provide the code of the psr interface packages. It provides implementations of the interfaces and so should provide the virtual packages representing implementations of the interface.
Providing the wrong package while also requiring it creates issues with Composer 2, because the solver will consider that installing the psr packages is not necessary as they are already provided.